+++ src/components/spartan/SDefinitionTerm/types.ts
@@ -1,5 +1,7 @@
+import type { ClassNameValue } from "tailwind-merge";
+
 export type TDefinitionTermProps = {
-    class?: string;
+    class?: ClassNameValue;
     labels?: string | string[];
     description?: string;
     oneline?: boolean;