+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
+
+## 2.4.0 - 2025-07-03
 ### Added
 - `laravel` prop to `SPaginator` component for Laravel pagination support.
 - `inertiaRouter` prop to `SPaginator` component for Inertia router integration.